flash_os/requirements.txt
psutil==5.9.1
pyserial==3.5
pyudev==0.23.2
